+++ packages/dialect-react-ui/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## [UNRELEASED]
 
+- fix: show solana name service domain without middle truncating
+
 ## [0.9.1] - 2022-06-15
 
 - fix: type error with initial address state of create thread page